==> notebooks/1_create_labels_dataset.ipynb <==
import os
import shutil
import urllib.request
import pandas as pd
from tqdm import tqdm
from time import sleep

# Load

wine_df = pd.read_pickle('../data/cleaned_dataset.gzip', compression='gzip')
print(wine_df.shape)

wine_df.tail(2)

raw_df = pd.read_pickle('../data/scraped/scraped_with_decs.pickle.gzip', compression='gzip')
print(raw_df.shape)

raw_df.tail(2)

# Merge

merged_df = raw_df.reset_index()\
              .rename(columns={'index':'file_index'})\
              .merge(right=wine_df, how='left', on='name')\
              .drop_duplicates(subset='name')
assert merged_df.shape[0] == raw_df.shape[0]

#wine_types = ['Red Wine', 'White Wine', 'Pink and Rosé', 'Sparkling & Champagne']
wine_types = merged_df['category_2'].value_counts(normalize=True)[:15].index
merged_df['origin_short'] = merged_df['origin'].str.split().str[-1]
origins = merged_df['origin_short'].value_counts()[:10].index

clean_df = merged_df[merged_df['category_2'].isin(wine_types)]
clean_df = clean_df[clean_df['origin_short'].isin(origins)]
print(clean_df.shape)

# Iterate

labels_to_keep = clean_df['file_index'].values
print(labels_to_keep.shape)

source_dir = 'D:/data/sorted/label/'
dest_dir = 'D:/data/sorted/labels_clean/'
5/0
not_found = []
for label in tqdm(labels_to_keep):
  filename = f"label_{label}.jpg"
  try:
    shutil.copy(
      src=os.path.join(source_dir, filename),
      dst=os.path.join(dest_dir, filename)
    )
  except:
    not_found.append(filename)

not_found.__len__()

# Create metadata

clean_df.head(2)

category_mapping = dict([(y,x+1) for x,y in enumerate(sorted(set(clean_df['category_2'])))])
clean_df['cat_2_int'] = [category_mapping[x] for x in clean_df['category_2']]

clean_df.drop_duplicates('cat_2_int')[['category_1','category_2','cat_2_int']]

directory = 'D:/data/sorted/labels_clean/'

label_metadata = {}
label_metadata['labels'] = []
for entry in tqdm(os.listdir(directory)):
  filename = os.path.split(entry)[-1]
  if filename.split('.')[1] != 'jpg':
    continue
  file_index = filename.split('_')[-1].split('.jpg')[0]
  assert file_index.isdigit()
  wine_row = clean_df[clean_df['file_index'] == int(file_index)]
  assert wine_row.shape[0] == 1
  label_metadata['labels'].append((
    filename, 
    int(wine_row['cat_2_int'].values[0])
    #wine_row['origin_short'].values[0]
  ))

# Dump to JSON
with open('D:/data/sorted/labels_clean/dataset.json', 'w') as outfile:
    json.dump(label_metadata, outfile)
