==> notebooks/1_scrape_data_v1.ipynb <==
from requests import get
from requests.exceptions import RequestException
from contextlib import closing
from bs4 import BeautifulSoup
import pandas as pd
import decimal as decimal


wine_request = get('https://www.wine.com/list/wine/7155/1')

wine_html = wine_request.text
# Then make the soup, specifying the "lxml" parser
house_soup = BeautifulSoup(wine_html,'lxml')
mydivs = house_soup.findAll("div", {"class": "prodItem_wrap"})

wineNames = []
wineImgs = []
winePrices = []
wineUrls = []
raw_htmls = []

for product in mydivs:
    wineNames.append(product.find('span', {'class' : 'prodItemInfo_name'}).text)
    wineImgs.append(str(product.find('img')).split('src=')[1].split('/>')[0].replace('"', ''))
    winePrices.append(decimal.Decimal(str(product.find('meta', {'itemprop' : 'price'})).split('content=')[1].split(" itemprop")[0].replace('"', '').replace('$','')))
    wineUrls.append(product.find('a', {'class' : 'prodItemImage_link'})['href'])
    raw_htmls.append(str(product))

def scrapeHTML()

def read_wine_page(wine_request):
    global house_soup
    # The .text returns the text from the request
    wine_html = wine_request.text
    # Then make the soup, specifying the "lxml" parser
    house_soup = BeautifulSoup(wine_html,'lxml')
    mydivs = house_soup.findAll("div", {"class": "prodItem_wrap"})

    wineNames = []
    wineImgs = []
    winePrices = []
    wineUrls = []
    raw_htmls = []
    
    for product in mydivs:
        wineNames.append(product.find('span', {'class' : 'prodItemInfo_name'}).text)
        wineImgs.append(str(product.find('img')).split('src=')[1].split('/>')[0].replace('"', ''))
        winePrices.append(decimal.Decimal(str(product.find('meta', {'itemprop' : 'price'})).split('content=')[1].split(" itemprop")[0].replace('"', '').replace('$','')))
        wineUrls.append(product.find('a', {'class' : 'prodItemImage_link'})['href'])
        raw_htmls.append(str(product))
    
    
    winePage_df = pd.DataFrame(data=[wineNames,
                                     wineImgs,
                                     winePrices,
                                     wineUrls,
                                     raw_htmls]).transpose()
    winePage_df.columns = ['name','image_path','price','url','raw_html']
    
    return winePage_df


# Run through page URLS and grab basic information
wine_page_ix = 1
while True:
    if wine_page_ix == 1:
        get_response = get(f'https://www.wine.com/list/wine/{wine_page_ix}')
        df_wine_master = read_wine_page(get_response)
        wine_page_ix += 1
        
    # Pull request and check if OK response
    get_response = get(f'https://www.wine.com/list/wine/7155/{wine_page_ix}')
    temp_df = read_wine_page(get_response)
    if get_response.status_code != 200 or temp_df.shape[0] < 20:
        print("Finished pulling data, found {} status code".format(get_response.status_code))
        break
    
    # Append results
    df_wine_master = pd.concat([df_wine_master,temp_df], axis=0)
    print("{}: df shape: {}".format(wine_page_ix, df_wine_master.shape))
    wine_page_ix += 1
    
df_wine_master.to_pickle('data/scraped/scrape.pickle')

rows_before = df_wine_master.shape[0]
df_wine_master_noDupe = df_wine_master.drop_duplicates(subset='name')
rows_after = df_wine_master_noDupe.shape[0]

print(f"Removed {rows_before - rows_after:,} duplicate rows")

# Get Image Data

df_wine_list = pd.read_pickle('data/scraped/scrape.pickle').reset_index(drop=True)
print(f"df_wine_list: {df_wine_list.shape}")

df_wine_list.head()

from PIL import Image
import requests
from io import BytesIO
import matplotlib.pyplot as plt
from tqdm import tqdm
from urllib.request import urlretrieve


image_dict = {}
counter = 0
for path in tqdm(df_wine_list['image_path']):
    path = "https://www.wine.com" + path
    try:
        urlretrieve(path, "images/labels/label_{:05d}.jpg".format(counter))
    except:
        continue
    image_dict[counter] = df_wine_list['name'][counter]
    counter += 1
        
pd.Series(image_dict).to_csv('images/labels/name_label_index.csv')

import os
import glob
TARGET_SIZE = 10 * 1024

files = glob.glob("images/labels/label_*.jpg")
for file in tqdm(files):
    if os.stat(file).st_size < TARGET_SIZE:
        os.remove(file)

import shutil

files = glob.glob("images/labels/label_*.jpg")
for file in tqdm(files):
    fileName = file.split("\\")[-1]
    shutil.move(file,f"data/scraped/labels/{fileName}")

# NO NEED

from PIL import Image, ImageOps
import glob
from tqdm import tqdm


desired_size = 512

raw_images = glob.glob('images/labels/*.jpg')
for raw_image in tqdm(raw_images):
    #print(raw_image)
    im = Image.open(raw_image)
    old_size = im.size
    ratio = float(desired_size)/max(old_size)
    new_size = tuple([int(x*ratio) for x in old_size])
    im = im.resize(new_size, Image.ANTIALIAS)
    new_im = Image.new("RGB", (desired_size, desired_size), color='white')
    new_im.paste(im, ((desired_size-new_size[0])//2,
                        (desired_size-new_size[1])//2))
    new_im.save("images/labels_padded/{}".format(raw_image.split('\\')[-1]))
