==> notebooks/1_scrape_labels.ipynb <==
import urllib.request
import pandas as pd
from tqdm import tqdm
from time import sleep

wine_df = pd.read_pickle('data/best_wine_df_ever.pickle.gzip', compression='gzip')

wine_df.head(3)

raw_df = pd.read_pickle('data/scraped/scraped_with_decs.pickle.gzip', compression='gzip')


raw_df.head(5)

raw_df.shape

raw_df.tail()

label_dir = '/mnt/d/data/wine/scraped_labels/'

for ix, wine in tqdm(raw_df.iterrows(), total=len(raw_df)):
    #print(wine['name'])
    #rint(wine['image_path'])
    sleep(1)
    try:
        urllib.request.urlretrieve(f"http://www.wine.com/{wine['image_path']}", label_dir+f"label_{ix}.jpg")
    except:
        pass
