==> notebooks/2a_clean_dataset.ipynb <==
import numpy as np
import pandas as pd
import six
import time
import os
from tqdm import tqdm
import pandas as pd
import bs4

import random
import string

# Load raw scraped dataset

scraped_data = pd.read_pickle('data/scraped/scrape.pickle.gz', compression='gzip')
print(scraped_data.shape)

df = scraped_data.reset_index()
df[df['url'].str.contains('Trapiche-Falling-Star-Ch')]

scraped_data.head()

# Load scraped descriptions

descriptions = pd.read_csv(
  "data/scraped/descriptions.csv.gz",
  sep='|',
  header=None,
  names=['URL_name', 'desc_raw'],
  encoding='latin1',
  compression='gzip')

descriptions['description'] = descriptions['desc_raw'].apply(
    lambda x: bs4.BeautifulSoup(str(x), 'lxml').get_text()
    )

descriptions.drop('desc_raw', axis=1, inplace=True)

print(descriptions.shape)

descriptions.head(3)

# Clean up descriptions

descriptions['description'] = descriptions['description'].apply(
  lambda x: x.lstrip().rstrip()
)


descriptions.iloc[1,:]

# Merge datasets

scraped_data.drop_duplicates(inplace=True)
descriptions.drop_duplicates(inplace=True)

combined = pd.merge(
    left=scraped_data,
    right=descriptions,
    how='left',
    left_on='url',
    right_on='URL_name'
    )

print(combined.shape)

combined.drop_duplicates(subset='name',inplace=True)
combined.reset_index(drop=True, inplace=True)
print(combined.shape)

# Save

combined.to_pickle('data/scraped/scraped_with_decs.pickle.gzip', compression='gzip')
