==> notebooks/2b_build_gpt2_dataset.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

scraped = pd.read_pickle('../data/cleaned_dataset.gzip', compression='gzip')
scraped.head(3)

scraped = scraped.replace(r'\\n',' ', regex=True) 

gpt2_txt = scraped.copy()

gpt2_txt['start_token'] = "<|startoftext|>"
gpt2_txt['end_token'] = "<|endoftext|>"
gpt2_txt['prompt_token'] = "[prompt]"
gpt2_txt['response_token'] = "[response]"
gpt2_txt['category_1_token'] = "[category_1]"
gpt2_txt['category_2_token'] = "[category_2]"
gpt2_txt['origin_token'] = "[origin]"
gpt2_txt['description_token'] = "[description]"

gpt2_txt = gpt2_txt[
  [
    'start_token', 
    'prompt_token', 
    'name', 
    'response_token',
    'category_1_token',
    'category_1',
    'category_2_token',
    'category_2',
    'origin_token',
    'origin',
    'description_token',
    'description',
    'end_token'
  ]
]

gpt2_txt.head(3)

# Quality cleanup

print(gpt2_txt.shape)
gpt2_txt_clean = gpt2_txt[~gpt2_txt['name'].str.lower().str.contains('gift')]
print(gpt2_txt_clean.shape)
gpt2_txt_clean = gpt2_txt_clean[~gpt2_txt_clean['description'].str.lower().str.contains('gift')]
print(gpt2_txt_clean.shape)


gpt2_txt_clean.head()

gpt2_txt_clean['origin'].value_counts()[:60]

# Data split and save

train, test = train_test_split(gpt2_txt_clean, test_size=0.2, random_state=0)
print(train.shape)
print(test.shape)

gpt2_txt_clean.to_csv(
    "data/scraped/name_desc_nlp_ready.txt", 
    sep=' ',
    index=False,
    header=None)

train.to_csv(
    "data/scraped/name_desc_nlp_ready_train.txt", 
    sep=' ',
    index=False,
    header=None)

test.to_csv(
    "data/scraped/name_desc_nlp_ready_test.txt", 
    sep=' ',
    index=False,
    header=None)
