==> notebooks/2c_build_master_dataset.ipynb <==
import pandas as pd
import bs4
import glob
import re
import sys
import tqdm

from ferment_wine import ferment_wine

DATA_DIR = 'E:/wine_pages/'

scraping_log = pd.read_csv(DATA_DIR+'log.txt', header=None, sep='\t')
scraping_log['response_code'] = scraping_log[1].str.split('<ResponseCode>:', expand=True)[1]
scraping_log['url'] = scraping_log[0].str.split('<URL>:', expand=True)[1]
scraping_log = scraping_log[['url','response_code']]

scraping_log.head()

scraping_log['response_code'].value_counts()

file_paths = glob.glob(DATA_DIR+'200*html')
print(f"Total 200 status pages found: {len(file_paths):,}")

import multiprocessing as mp

p = mp.Pool(processes=16)
df_list = p.map(ferment_wine, file_paths)

errs = wine_df[wine_df['category_1'].isnull()]
print(errs.shape)

wine_df.isnull().sum()

wine_df.sample(50)

wine_df = pd.concat(df_list)
print(wine_df.shape)

wine_df.to_pickle('data/best_wine_df_ever.pickle.gzip', compression='gzip')

# Scratch

print(bad_wines[0])

wine_name = soup.find('h1',attrs={'itemprop':'name'}).text
wine_name

wine_description = soup.find('div', attrs={'class':'viewMoreModule_text'})
print(wine_description)
if len(wine_description.find_all('p')) > 0:
  fixed_description = '\n'.join(str(tag.text) for tag in wine_description.find_all('p'))
else:
  fixed_description = wine_description.text
print("\n",fixed_description)

wine_price = soup.find('div', attrs={'class','productPrice'})
if wine_price is None:
  wine_price = soup.find('meta', attrs={'itemprop':'price'})['content']
else:
  wine_price = wine_price.find('meta', {'itemprop':'price'})['content']
print(wine_price)

soup.find('meta', attrs={'itemprop':'price'})['content']

soup.find('div', attrs={'class','productPrice'})

wine_category_1 = soup.find("ul", attrs={"class":"prodAttr"}).find("li")['title']
print(wine_category_1)

wine_category_2 = soup.find("span", {"class":"prodItemInfo_varietal"}).text
print(wine_category_2)

wine_origin = soup.find("span", {"class":"prodItemInfo_originText"}).text
print(wine_origin)

wine_size_value = soup.find("span", attrs={"class":"prodAlcoholVolume_text"}).text
print(wine_size_value)

wine_size_units = re.findall("[a-zA-Z]+", soup.find("span", attrs={"class":"prodAlcoholVolume"}).text)[0]
print(wine_size_units)

wine_abv = soup.find("span", attrs={"class":"prodAlcoholPercent_percent"}).text
print(wine_abv)

---------------------

# OLD

raw_df = pd.read_pickle('data/scraped/scraped_with_decs.pickle.gzip', compression='gzip')
print(raw_df.shape)

raw_df.head(3)

master_df = raw_df.copy()

for ix,row in master_df.iterrows():
  raw_html = row['raw_html']
  soup = bs4.BeautifulSoup(raw_html)
  if ix == 2:
    break

wine_category_1 = soup.find("ul", attrs={"class":"prodAttr"}).find("li")['title']
print(wine_category_1)

wine_category_2 = soup.find("span", {"class":"prodItemInfo_varietal"}).text
print(wine_category_2)

wine_origin = soup.find("span", {"class":"prodItemInfo_originText"}).text
print(wine_origin)

wine_size_value = soup.find("span", attrs={"class":"prodAlcoholVolume_text"}).text
print(wine_size_value)

wine_size_units = re.findall("[a-zA-Z]+", soup.find("span", attrs={"class":"prodAlcoholVolume"}).text)[0]
print(wine_size_units)

wine_abv = soup.find("span", attrs={"class":"prodAlcoholPercent_percent"}).text
print(wine_abv)

soup.find("span", attrs={"class":"prodAlcoholPercent_percent"}).text


import re

re.findall("[a-zA-Z]+", soup.find("span", attrs={"class":"prodAlcoholVolume"}).text)

soup.find("span", attrs={"class":"prodAlcoholVolume_text"}).text

soup.find("span", attrs={"class":"prodAlcoholVolume"}).text.strip()

url = "https://www.wine.com/product/laurent-perrier-cuvee-rose/10521"
html = requests.get(url)

str(html)

import requests

url = "https://www.wine.com/product/laurent-perrier-cuvee-rose/10521"
html = requests.get(url).content

soup = bs4.BeautifulSoup(html, "html.parser")

soup.find("div", attrs={"class":"prodAlcoholPercent"})

old_scrape = pd.read_pickle('data/scraped/scrape.pickle.gz', compression='gzip')
old_scrape.head(4)

old_scrape.iloc[2,:]['url']

bs4.BeautifulSoup(old_scrape.iloc[2,:]['raw_html'])

soup
