==> notebooks/4b_generate_descriptions.ipynb <==
from collections import OrderedDict 
import re

import pandas as pd
import pickle
import torch
import tqdm
import transformers

print(torch.__version__)
print(transformers.__version__)

# Load tokenizer and fine-tuned model

TOKENIZER_PATH = "data/gpt2_runs/tokenizers/gpt2_large"
MODEL_PATH = "data/gpt2_runs/v2/step_290000/"

tokenizer = transformers.GPT2Tokenizer.from_pretrained(TOKENIZER_PATH)
model = transformers.AutoModelForCausalLM.from_pretrained(
  MODEL_PATH,
  pad_token_id=tokenizer.eos_token_id
)

model = model.to('cuda:1')

# Greedy Output

input_prompt = "<|startoftext|> [prompt] Walter Hansel Cuvee Alyce Chardonnay 2015 [response] "
input_ids = tokenizer.encode(input_prompt, return_tensors='pt').to('cuda:1')
print("Tokens: ", input_ids)

greedy_output = model.generate(input_ids, max_length=200)
print(tokenizer.decode(greedy_output[0]))

# Beam Output

input_prompt = "<|startoftext|> [prompt] Walter Hansel Cuvee Alyce Chardonnay 2015 [response] "
input_ids = tokenizer.encode(input_prompt, return_tensors='pt').to('cuda:1')

# activate beam search and early_stopping
beam_outputs = model.generate(
  input_ids, 
  max_length=200, 
  num_beams=5, 
  no_repeat_ngram_size=2,
  num_return_sequences=3, 
  early_stopping=True
)

print("Output:\n" + 80 * '-')
for i, beam_output in enumerate(beam_outputs):
  print("="*20)
  print("{}: {}".format(i, tokenizer.decode(beam_output, skip_special_tokens=True)))
  print('\n')

# Random Sampling Output

input_prompt = "<|startoftext|> " + "[prompt] " + "Vendange Chardonnay " + "[response] "
input_ids = tokenizer.encode(input_prompt, return_tensors='pt')
print(input_ids)

# Send to GPU
model.to('cuda:1')
input_ids = input_ids.to('cuda:1')

sample_output = model.generate(
    input_ids, 
    do_sample=True, 
    max_length=250, 
    top_p=0.8,
    top_k=200,
    temperature=0.9,
    eos_token_id=tokenizer.eos_token_id,
    bos_token_id=tokenizer.eos_token_id,
    early_stopping=True
)

print("Output:\n" + 100 * '-')
print("="*20)
print(tokenizer.decode(sample_output[0], skip_special_tokens=True))
print('\n')

# Generate descriptions on fake wine names

names_path = 'data/fake/fake_names_12184_2020-11-19.pickle'
with open(names_path, 'rb',) as file:
    fake_names = pickle.load(file)

fake_names.head(3)

# Send to GPU
model.to('cuda:1')
#input_ids = input_ids.to('cuda:1')

generated_descriptions = OrderedDict()
for fake_name in tqdm.tqdm(fake_names[:1000]):
  #print(f"Name: {fake_name}")
  
  # Create token from fake wine name
  try:
    input_ids = tokenizer.encode(
      text=("<|startoftext|> [prompt] " + fake_name + " " + "[response] "), 
      return_tensors='pt'
    ).to('cuda:1')
  
    # Generate a fake description based on the name
    model_output = model.generate(
      input_ids, 
      do_sample=True, 
      max_length=300,
      min_length=80,
      top_p=0.8,
      top_k=200,
      temperature=0.9,
      eos_token_id=tokenizer.eos_token_id,
      bos_token_id=tokenizer.bos_token_id,
      early_stopping=True
    )

    generated_descriptions[fake_name] = tokenizer.decode(
      token_ids=model_output[0], 
      skip_special_tokens=True
    )
    
  except:
    continue

wine_df = pd.DataFrame.from_dict(generated_descriptions.items())
wine_df.columns = ['name', 'response']
print(wine_df.shape)

wine_df['category_1'] = wine_df['response'].str.split('\[category_1\]').str[1].str.split('\[category_2\]').str[0]
wine_df['category_2'] = wine_df['response'].str.split('\[category_2\]').str[1].str.split('\[origin\]').str[0]
wine_df['origin'] = wine_df['response'].str.split('\[origin\]').str[1].str.split('\[description\]').str[0]
wine_df['description'] = wine_df['response'].str.split('\[description\]').str[1]
#wine_df['description'] = wine_df['description'].str.strip()
#wine_df['description'] = wine_df['description'].str.strip('"')

#wine_df = wine_df[wine_df['description'].str.len() > 100]
wine_df = wine_df.applymap(str)\
            .applymap(lambda x: x.replace('"', ''))\
            .drop(['response'], axis=1)

print(wine_df.shape)
wine_df.head()

wine_df.to_csv("data/fake/descriptions/gpt2_desc_v2_20210127.csv")
