==> notebooks/4b_v2_generate_gpt2_descriptions.ipynb <==
!pip install wandb; git clone https://github.com/huggingface/transformers; cd transformers; pip install .

import os
import pickle
from tqdm import tqdm

import pandas as pd
import torch
import transformers
from tqdm import tqdm


tokenizer_path = 'drive/MyDrive/data/wine/gpt2_large/'
tokenizer = transformers.GPT2Tokenizer.from_pretrained(tokenizer_path)
print("Loaded tokenizer")

names_path = 'drive/MyDrive/data/wine/fake_names_12184_2020-11-19.pickle'
with open(names_path, 'rb',) as file:
    fake_wine_names = pickle.load(file)

fake_wine_names[:5]

model_path = 'drive/MyDrive/data/wine/checkpoints/step_95000_loss_0.7529/'
model = transformers.AutoModelForCausalLM.from_pretrained(model_path)
model = model.to('cuda')
print("Loaded model")

tokenizer.encode('<|endoftext|>')

generated_descriptions = {}
for fake_name in tqdm(fake_wine_names):
  # Create token from fake wine name
  input_ids = tokenizer.encode(
    text=("<|startoftext|> [prompt] " + fake_name + " " + "[response] "),
    return_tensors='pt'
  ).to('cuda')

  # Generate a fake description based on the name
  model_output = model.generate(
    input_ids, 
    do_sample=True, 
    max_length=250, 
    top_p=0.8,
    top_k=200,
    temperature=0.9,
    eos_token_id=50257,
    bos_token_id=50256,
    pad_token_id=50256,
    early_stopping=True
  )
  
  generated_descriptions[fake_name] = tokenizer.decode(
    token_ids=model_output[0], 
    skip_special_tokens=True
  )

for name in generated_descriptions.keys():
  print(generated_descriptions[name])
  break

descriptions_df = pd.DataFrame.from_dict(generated_descriptions, orient='index')
descriptions_df.columns = ['gpt2_description']
descriptions_df.to_csv('drive/MyDrive/data/wine/gpt2_descriptions.csv')

descriptions_df

# Scratch

import pandas as pd

data_raw = pd.read_csv("../data/fake/gpt2_xl/generated_desc_201.csv")
print(data_raw.shape)
data_raw.head(3)

data = data_raw[data_raw['error_msg'] == 'None']
data = data[['name','output']].replace('"', '', regex=True)

print(data.shape)
data.head(3)

category_1 = data['output']\
    .str.split('\[category_1\]', expand=True)[1]\
    .str.split('\[category_2\]', expand=True)[0]

category_2 = data['output']\
    .str.split('\[category_2]', expand=True)[1]\
    .str.split('\[origin]', expand=True)[0]

origin = data['output']\
    .str.split('\[origin]', expand=True)[1]\
    .str.split(' from ', expand=True)[1]\
    .str.split('\[description\]', expand=True)[0]

description = data['output']\
    .str.split('\[description]', expand=True)[1]

data_clean = pd.DataFrame({
  'name': data['name'],
  'category_1': category_1,
  'category_2': category_2,
  'origin': origin,
  'description': description
})

data_clean.head(3)

category_1_valids = ['Red Wine', 'White Wine', 'Pink and Rosé', 'Sparkling & Champagne']
category_2_valids = ['Chardonnay', 'Pinot Noir', 'Cabernet Sauvignon', 
                     'Bordeaux Red Blends', 'Other Red Blends', 'Syrah/Shiraz',
                     'Sauvignon Blanc', 'Merlot', 'Sangiovese', 'Rhone Red Blends', 
                     'Zinfandel', 'Riesling', 'Rosé', 'Pinot Gris/Grigio', 
                     'Other White Blends', 'Tempranillo', 'Nebbiolo', 'Malbec']

data_clean = data_clean.applymap(lambda x: x.strip() if isinstance(x, str) else x)
data_clean = data_clean[data_clean['category_1'].isin(category_1_valids)]
data_clean = data_clean[data_clean['category_2'].isin(category_2_valids)]


data_clean
